backend/healthy/requirements.txt
djangorestframework
django-cors-headers
pandas
sklearn
matplotlib